xterm 238. includes fix for CVE-2008-2383.

+++ app/xterm/charclass.c
@@ -1,4 +1,4 @@
-/* $XTermId: charclass.c,v 1.20 2008/01/31 01:07:23 tom Exp $ */
+/* $XTermId: charclass.c,v 1.21 2008/12/30 17:35:09 tom Exp $ */
 
 /*
  * Compact and efficient reimplementation of the
@@ -51,7 +51,7 @@ SetCharacterClassRange(int low, int high, int value)
     /* make sure we have at least one free entry left at table end */
     if (classtab[0].last > classtab[0].cclass - 2) {
 	classtab[0].cclass += 5 + classtab[0].cclass / 4;
-	classtab = TypeRealloc(struct classentry, classtab[0].cclass, classtab);
+	classtab = TypeRealloc(struct classentry, (unsigned) classtab[0].cclass, classtab);
 	if (!classtab)
 	    abort();
     }
